requirements.txt
boto3==1.35.77
botocore==1.35.77
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
numpy==2.0.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.17.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
